[DOK-225] style : mypage 파일 구조변경
