+++ src/pages/User/MyPage/Inquiry/InquiryHistoryPage/InquiryHistoryPage.js
@@ -1,0 +1,13 @@
+import React from "react";
+import SidebarLayout from "components/templete/SidebarLayout"
+import InquiryHistoryBody from "./InquiryHistoryBody";
+
+const InquiryHistory = () => {
+  return (
+    <SidebarLayout>
+      <InquiryHistoryBody />
+    </SidebarLayout>
+  );
+};
+
+export default InquiryHistory;
