+++ src/pages/User/MyPage/Inquiry/InquiryWritePage/InquiryWritePage.js
@@ -1,0 +1,14 @@
+import React from "react";
+import SidebarLayout from "../../../components/templates/SidebarLayout";
+import InquiryWriteBody from "./InquiryWriteBody";
+
+
+const Inquiry = () => {
+  return (
+    <SidebarLayout>
+      <InquiryWriteBody />
+    </SidebarLayout>
+  );
+};
+
+export default Inquiry;
