+++ src/pages/User/MyPage/MyInfoPage/MyInfoPage.js
@@ -1,0 +1,13 @@
+import React from "react";
+import SidebarLayout from "components/templete/SidebarLayout"
+import MyInfoBody from "./MyInfoBody";
+
+const MyInfo = () => {
+  return (
+    <SidebarLayout>
+      <MyInfoBody />
+    </SidebarLayout>
+  );
+};
+
+export default MyInfo;
