+++ src/pages/User/MyPage/Order/OrderListPage/OrderDetailPage.js
@@ -1,0 +1,13 @@
+import React from "react";
+import SidebarLayout from "components/layout/SidebarLayout";
+import OrderDetailBody from "./OrderDetailBody";
+
+const OrderDetailPage = () => {
+  return (
+    <SidebarLayout>
+      <OrderDetailBody />
+    </SidebarLayout>
+  );
+};
+
+export default OrderDetailPage;
