+++ src/pages/User/MyPage/Order/OrderListPage/OrderListPage.js
@@ -1,0 +1,15 @@
+import React from "react";
+import SidebarLayout from "components/templete/SidebarLayout";
+import OrderListBody from "./OrderListBody";
+
+
+const OrderList = () => {
+
+  return (
+    <SidebarLayout>
+        <OrderListBody />
+    </SidebarLayout>
+  );
+};
+
+export default OrderList;
