+++ src/pages/User/MyPage/Order/RefundPage/RefundPage.js
@@ -1,0 +1,12 @@
+import React from "react";
+import SidebarLayout from "components/templete/SidebarLayout"
+import RefundBody from "./RefundBody";
+const Refund = () => {
+  return (
+    <SidebarLayout>
+      <RefundBody />
+    </SidebarLayout>
+  );
+};
+
+export default Refund;
