+++ src/pages/User/MyPage/Review/ReviewPage/ReviewPage.js
@@ -1,0 +1,13 @@
+import React from "react";
+import SidebarLayout from "components/templete/SidebarLayout"
+import MyReviewBody from "./ReviewBody";
+
+const Review = () => {
+  return (
+    <SidebarLayout>
+      <MyReviewBody />
+    </SidebarLayout>
+  );
+};
+
+export default Review;
